Add files via upload
This report describe the amount of CRs according to the Maintenance Status.
The aim of this report is to identify personal blind spots and wrong assumptions from the process facilitator and make the documentation more accurate and useful.
It makes you understand what are the CR that needs to be mantain during the project. 
Gain/benefit: Creates an overview of the Maintenance CR & IAL e.g. Process facilitator can see the GAPS in any type of status like new or rejected. Process facilitator can see also according to the source, their correspond date as DA, Go live date and UAT Date, etc. Even Process facilitator can also see for each of the CR if it is already created or not. 
Justification: Process facilitator follow up status of maintenance IAL and CRs, also who has been assigned to, the priority and the requesting Team
I used different tables, create relationship between each other, slicers to filter by priority, status. cards to know the number of CR that need to be mantain.
The matrix was used to see the CRs by teams according to the last years.

RFC existing=True when RFC_ID have entries and RFC existing= False when RFC_ID is in blank.  
Remove HTML:
Let
Source = if HTML = null then
        ""
    else
        Text.From(HTML),
    SplitAny = Text.SplitAny(Source,"<>"),
    ListAlternate = List.Alternate(SplitAny,1,1,1),
    ListSelect = List.Select(ListAlternate, each _<>""),
    TextCombine = Text.Combine(ListSelect, ""),
    TextTrim = Text.Trim(TextCombine)
in
    TextTrim
in
    TextFromHtml