Status report
WHAT - This is a status report for the Release flow.
Scope: The report provides an opportunity to get an overall status of the release and at the same time it will be possible to see a more detailed status for specific areas.
For example:
-  Release level
-   Team/Entity level
-    Task level
-   Comments (Dynamic Task List)
Data Scope (If applicable)
The sources are from (CR, IAL, RFC, CD, SharePoint list and Dynamic List)
Permissions needed (e.g. access to Dynamic List, SharePoint, Power Bi, etc.) (RISK) 
Category: Reports/Dashboards, Data tools
Duration: Multiple Cross Release
Adaptability/Reusability: Not reusability. It is specific for PMO Methodology

WHY –We want to distribute a uniform status across releases 
Gain/benefit: e.g. work-Hours saved, increasing data quality, no need to create another Management Dashboard.
Justification: The reporting for a release has been individual agreed between the Release Lead and the Release Coordinator, so it has never been exactly the same information in the same format.
WHO – consists of stakeholders, there are many who are working on multiple releases, and therefore it will be easier to relate to the status if it is the same for all releases. 
Ownership: Release coordinator.
Who will be directly affected by changes?
Release Leads and Co- Leads
WHEN – Consists of timeframe related to:
During the whole release which means from the start (IAL) until the end (CD).

HOW – Through Power Bi.
There is dependency data base and SharePoint List. 
There is also data related with the Dynamic List in order to obtain the comments. 
This dashboards can be distributed by Team.



Maintenance level:
Maintained by the Release Coordinator. 
Volatility of context (RISK) : No. It is actually quite stable. 
The structure of the report that contains the IAL, CR, RFC and CD don’t change